pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "HelixPipe" # 给您的项目起个名字
version = "0.1.0"
description = "A reproducible, config-driven data engine for building heterogeneous knowledge graphs for drug discovery."
requires-python = ">=3.9"
authors = [
    { name = "Chujia Zhu", email = "wozonet12345@gmail.com" },
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]

# --- [核心部分在这里] ---
# 列出你项目运行时的【核心】依赖
dependencies = [
    # --- 核心数据处理 ---
    "pandas>=1.5.0",
    "numpy>=1.21.0",
    "PyYAML>=6.0",
    "joblib",

    # --- 配置管理 ---
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",

    # --- 生物/化学信息学 ---
    "rdkit>=2022.09.1",    # 假设你使用的是 rdkit-pypi，这是最常见的
    "bioservices",
    "requests",

    # --- 机器学习 & GNN (核心依赖) ---
    "torch>=2.0.0",
    "torch-geometric>=2.3.0",
    "faiss-cpu>=1.7.0",   # 或者 "faiss-gpu"，取决于你的主要目标环境

    # --- 其他工具 ---
    "tqdm",
]

[project.optional-dependencies]
# 依赖项用于开发、测试和质量检查
dev = [
    "pytest",
    "mypy",
    "black",
    "isort",
    "pre-commit",  # 强烈推荐，用于自动化代码检查
    "pandas-stubs",
    "types-PyYAML",
    "types-requests",
    "types-tqdm",
]

# 依赖项用于构建文档
docs = [
    "mkdocs",
    "mkdocs-material",
    "mkdocstrings[python]",
]

# 一个方便的分组，包含所有非核心的依赖项
all = [
    "helixpipe[dev]",
    "helixpipe[docs]",
]

[tool.setuptools.packages.find]
where = ["src"]  # 告诉setuptools，包的根目录是 'src'
include = ["helix*"]  # 包含所有以 'helixpipe' 开头的包
exclude = []  # 可以排除特定的子包
[tool.setuptools.package-dir]
# 这一部分将包名映射到实际的目录
helixpipe = "src/helixpipe"
helixlib = "src/helixlib"
[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501",]


[tool.pyright]
# 定义分析范围
include = ["src", "tests", "analysis"]
# 定义导入路径
extraPaths = ["src"]
# 默认检查模式
typeCheckingMode = "standard"

# 定义 'src' 环境的严格规则
[[tool.pyright.executionEnvironments]]
root = "src"
typeCheckingMode = "strict"

# 定义 'tests' 和 'analysis' 环境的宽松规则
[[tool.pyright.executionEnvironments]]
root = "tests"
reportArgumentType = "none" # 允许 Mock 对象

[[tool.pyright.executionEnvironments]]
root = "analysis"
reportArgumentType = "none"


# ==========================================================
# MyPy Configuration - For CI/CD and Batch Analysis
# ==========================================================
[tool.mypy]
# 定义导入路径 (与 extraPaths 对应)
mypy_path = "src/"

# --- 核心严格选项 (与 Pyright 'strict' 模式对齐) ---
check_untyped_defs = true
disallow_any_unimported = true
disallow_untyped_defs = true
no_implicit_optional = true
show_error_codes = true
strict_equality = true
warn_redundant_casts = true
warn_return_any = true
warn_unused_configs = true
warn_unused_ignores = true



# --- 忽略第三方库的导入错误 (与 Pyright 的默认行为对齐) ---
[[tool.mypy.overrides]]
module = [
    "pandas.*", "numpy.*", "torch.*", "torch_geometric.*", "omegaconf.*", 
    "hydra.*", "rdkit.*", "tqdm.*", "sklearn.*", "research_template.*", "joblib.*"
]
ignore_missing_imports = true

# --- 【核心】为 'tests' 和 'analysis' 目录定义宽松规则 ---
# (MyPy 使用 per-file 规则，而不是像 Pyright 那样的环境)
[[tool.mypy.overrides]]
# 匹配 tests/ 和 analysis/ 目录下的所有文件
module = ["tests.*", "analysis.*"]
# 关闭“参数类型不匹配”的检查
disable_error_code = "arg-type"